--- modules/BookList.js
@@ -1,10 +1,0 @@
-export let books = [
-  {
-    author: 'ipsum',
-    title: 'Testyy',
-  },
-  {
-    author: 'lorem',
-    title: 'Testeroo',
-  },
-];
