+++ modules/RemoveBook.js
@@ -1,5 +1,5 @@
 import { setLocalStore } from './localStorage.js';
-import { books } from './BookList.js';
+import { books } from './start.js';
 
 export const removeBook = (index) => {
   books.splice(index, 1);
